research/trusted-sources/requirements.txt
# Trusted Sources Research Prototype Dependencies

# Web scraping and content extraction
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1

# Data processing
sqlite3  # Built-in with Python

# CLI interface
argparse  # Built-in with Python

# Logging
logging  # Built-in with Python

# Text processing
difflib  # Built-in with Python
re  # Built-in with Python

# Optional: For future Norwegian language processing
# langdetect>=1.0.9